# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.createTaskRequest:
    properties:
      description:
        maxLength: 2000
        type: string
      due_date:
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - due_date
    - title
    type: object
  handler.errorResponseBody:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
    type: object
  handler.taskResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  handler.updateTaskRequest:
    properties:
      description:
        maxLength: 2000
        type: string
      due_date:
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - due_date
    - title
    type: object
host: localhost:8080
info:
  contact: {}
  description: A small task management REST API built for learning Clean Architecture
    in Go.
  title: Task API
  version: "1.0"
paths:
  /tasks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.taskResponse'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponseBody'
      summary: List all tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Creates a new task with the given title, description, and due date
      parameters:
      - description: Task payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.taskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponseBody'
      summary: Create a task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponseBody'
      summary: Delete a task
      tags:
      - tasks
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.taskResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponseBody'
      summary: Get a task by ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Updates an existing task's title, description, and due date
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated task payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.taskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponseBody'
      summary: Update a task
      tags:
      - tasks
  /tasks/{id}/complete:
    post:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.taskResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponseBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponseBody'
      summary: Mark a task as completed
      tags:
      - tasks
swagger: "2.0"