# internal/handler/task_handler.go
// internal/handler/task_handler.go
package handler

import (
	"encoding/json"
	"net/http"

	"github.com/az/task-api/internal/apperror"
	"github.com/az/task-api/internal/usecase"
	"github.com/go-chi/chi/v5"
)

type TaskHandler struct {
	usecase *usecase.TaskUsecase
}

func NewTaskHandler(u *usecase.TaskUsecase) *TaskHandler {
	return &TaskHandler{usecase: u}
}

// Create godoc
// @Summary      Create a task
// @Description  Creates a new task with the given title, description, and due date
// @Tags         tasks
// @Accept       json
// @Produce      json
// @Param        request body createTaskRequest true "Task payload"
// @Success      201 {object} taskResponse
// @Failure      400 {object} errorResponseBody
// @Failure      500 {object} errorResponseBody
// @Router       /tasks [post]
func (h *TaskHandler) Create(w http.ResponseWriter, r *http.Request) {
	var req createTaskRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "invalid request body", http.StatusBadRequest)
		return
	}

	if err := validateStruct(req); err != nil {
		writeError(w, apperror.Validation(err.Error()))
		return
	}

	task, err := h.usecase.CreateTask(r.Context(), req.Title, req.Description, req.DueDate)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError) // temporary — fixed in Step 10
		return
	}

	writeJSON(w, http.StatusCreated, toTaskResponse(task))
}

// Get godoc
// @Summary      Get a task by ID
// @Tags         tasks
// @Produce      json
// @Param        id path string true "Task ID"
// @Success      200 {object} taskResponse
// @Failure      404 {object} errorResponseBody
// @Failure      500 {object} errorResponseBody
// @Router       /tasks/{id} [get]
func (h *TaskHandler) Get(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")

	task, err := h.usecase.GetTask(r.Context(), id)
	if err != nil {
		writeError(w, err)
		return
	}

	writeJSON(w, http.StatusOK, toTaskResponse(task))
}

// List godoc
// @Summary      List all tasks
// @Tags         tasks
// @Produce      json
// @Success      200 {array} taskResponse
// @Failure      404 {object} errorResponseBody
// @Failure      500 {object} errorResponseBody
// @Router       /tasks [get]
func (h *TaskHandler) List(w http.ResponseWriter, r *http.Request) {
	tasks, err := h.usecase.ListTasks(r.Context())
	if err != nil {
		writeError(w, err)
		return
	}

	resp := make([]taskResponse, len(tasks))
	for i, t := range tasks {
		resp[i] = toTaskResponse(t)
	}
	writeJSON(w, http.StatusOK, resp)
}

// Update godoc
// @Summary      Update a task
// @Description  Updates an existing task's title, description, and due date
// @Tags         tasks
// @Accept       json
// @Produce      json
// @Param        id path string true "Task ID"
// @Param        request body updateTaskRequest true "Updated task payload"
// @Success      200 {object} taskResponse
// @Failure      400 {object} errorResponseBody
// @Failure      404 {object} errorResponseBody
// @Failure      500 {object} errorResponseBody
// @Router       /tasks/{id} [put]
func (h *TaskHandler) Update(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")

	var req updateTaskRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeError(w, apperror.Validation("invalid request body"))
		return
	}

	if err := validateStruct(req); err != nil {
		writeError(w, apperror.Validation(err.Error()))
		return
	}

	task, err := h.usecase.UpdateTask(r.Context(), id, req.Title, req.Description, req.DueDate)
	if err != nil {
		writeError(w, err)
		return
	}

	writeJSON(w, http.StatusOK, toTaskResponse(task))
}

// Delete godoc
// @Summary      Delete a task
// @Tags         tasks
// @Param        id path string true "Task ID"
// @Success      204 "No Content"
// @Failure      404 {object} errorResponseBody
// @Failure      500 {object} errorResponseBody
// @Router       /tasks/{id} [delete]
func (h *TaskHandler) Delete(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")

	if err := h.usecase.DeleteTask(r.Context(), id); err != nil {
		writeError(w, err)
		return
	}

	w.WriteHeader(http.StatusNoContent)
}

// Complete godoc
// @Summary      Mark a task as completed
// @Tags         tasks
// @Produce      json
// @Param        id path string true "Task ID"
// @Success      200 {object} taskResponse
// @Failure      404 {object} errorResponseBody
// @Failure      409 {object} errorResponseBody
// @Failure      500 {object} errorResponseBody
// @Router       /tasks/{id}/complete [post]
func (h *TaskHandler) Complete(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")

	task, err := h.usecase.CompleteTask(r.Context(), id)
	if err != nil {
		writeError(w, err)
		return
	}

	writeJSON(w, http.StatusOK, toTaskResponse(task))
}

func writeJSON(w http.ResponseWriter, status int, v interface{}) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	json.NewEncoder(w).Encode(v)
}